+++ INTERX/tasks/cache_header_check.go
@@ -9,6 +9,7 @@ import (
 
 	common "github.com/KiraCore/sekai/INTERX/common"
 	"github.com/KiraCore/sekai/INTERX/config"
+	"github.com/KiraCore/sekai/INTERX/global"
 )
 
 // CacheHeaderCheck is a function to check cache headers if it's expired.
@@ -41,14 +42,14 @@ func CacheHeaderCheck(rpcAddr string, isLog bool) {
 						common.GetLogger().Info("[cache] Deleting file: ", path)
 					}
 
-					common.Mutex.Lock()
+					global.Mutex.Lock()
 					// check if file or path exists
 					if _, err := os.Stat(path); os.IsNotExist(err) {
-						common.Mutex.Unlock()
+						global.Mutex.Unlock()
 						return nil
 					}
 					err := os.Remove(path)
-					common.Mutex.Unlock()
+					global.Mutex.Unlock()
 
 					if err != nil {
 						if isLog {