+++ INTERX/types/main.go
@@ -132,6 +132,36 @@ type InterxRequest struct {
 	Params   []byte `json:"params"`
 }
 
+type QueryValidator struct {
+	Address    string `protobuf:"bytes,1,opt,name=address,proto3" json:"address,omitempty"`
+	Valkey     string `protobuf:"bytes,2,opt,name=valkey,proto3" json:"valkey,omitempty"`
+	Pubkey     string `protobuf:"bytes,3,opt,name=pubkey,proto3" json:"pubkey,omitempty"`
+	Proposer   string `protobuf:"bytes,4,opt,name=proposer,proto3" json:"proposer,omitempty"`
+	Moniker    string `protobuf:"bytes,5,opt,name=moniker,proto3" json:"moniker,omitempty"`
+	Website    string `protobuf:"bytes,6,opt,name=website,proto3" json:"website,omitempty"`
+	Social     string `protobuf:"bytes,7,opt,name=social,proto3" json:"social,omitempty"`
+	Identity   string `protobuf:"bytes,8,opt,name=identity,proto3" json:"identity,omitempty"`
+	Commission string `protobuf:"bytes,9,opt,name=commission,proto3" json:"commission,omitempty"`
+	Status     string `protobuf:"bytes,10,opt,name=status,proto3" json:"status,omitempty"`
+	Rank       int64  `protobuf:"varint,11,opt,name=rank,proto3" json:"rank,omitempty"`
+	Streak     int64  `protobuf:"varint,12,opt,name=streak,proto3" json:"streak,omitempty"`
+	Mischance  int64  `protobuf:"varint,13,opt,name=mischance,proto3" json:"mischance,omitempty"`
+}
+
+type AllValidators struct {
+	Status struct {
+		NetworkStopped     bool `json:"network_stopped"`
+		ActiveValidators   int  `json:"active_validators"`
+		PausedValidators   int  `json:"paused_validators"`
+		InactiveValidators int  `json:"inactive_validators"`
+		JailedValidators   int  `json:"jailed_validators"`
+		TotalValidators    int  `json:"total_validators"`
+		TotalWaiting       int  `json:"total_waiting"`
+	} `json:"status"`
+	Waiting    []string         `json:"waiting"`
+	Validators []QueryValidator `json:"validators"`
+}
+
 const (
 	// GET is a constant to refer GET HTTP Method
 	GET string = "GET"