+++ CHANGELOG.md
@@ -1,5 +1,10 @@
 # Changelog
 
+## [v0.1.17.1] - 02.25.2021
+### Added
+- GRPC query for proposals
+- GRPC query for validators + validator_signing_infos
+
 ## [v0.1.17] - 02.16.2021
 ### Fixed
 - Problem with ClaimValidator and PubKey encoding due to protocol buff Any type.